refactor: constant 패키지의 상수 클래스 수정 공통적으로 사용되는 상수만 Constant클래스로 남겨두고 나머지 상수는 사용되는 클래스 내부에 넣었습니다.
