+++ src/main/java/christmas/model/Order.java
@@ -1,10 +1,5 @@
 package christmas.model;
 
-import static christmas.util.constant.OrderManagerConst.INDEX_OF_AMOUNT;
-import static christmas.util.constant.OrderManagerConst.INDEX_OF_MENU;
-import static christmas.util.constant.OrderManagerConst.MENU_AND_AMOUNT_DELIMITER;
-import static christmas.util.constant.OrderManagerConst.ORDER_DELIMITER;
-
 import christmas.util.Menu;
 import christmas.validator.DomainValidator;
 import java.util.ArrayList;
@@ -15,6 +10,10 @@ import java.util.Map;
 import java.util.stream.Collectors;
 
 public class Order {
+    private static final String ORDER_DELIMITER = ",";
+    private static final String MENU_AND_AMOUNT_DELIMITER = "-";
+    private static final int INDEX_OF_MENU = 0;
+    private static final int INDEX_OF_AMOUNT = 1;
     private final String userInputOrder;
     private Map<Menu, Integer> order;
