+++ src/main/java/christmas/util/constant/DiscountConstant.java
@@ -1,0 +1,5 @@
+package christmas.util.constant;
+
+public class DiscountConstant {
+    public static final int NON_DISCOUNT_COST = 0;
+}
