--- src/main/java/christmas/util/constant/DiscountEventConst.java
@@ -1,9 +1,0 @@
-package christmas.util.constant;
-
-public class DiscountEventConst {
-    public static final int EVENT_YEAR = 2023;
-    public static final int EVENT_MONTH = 12;
-    public static final int EVENT_START_DATE = 1;
-    public static final int CHRISTMAS_DATE = 25;
-    public static final int NON_DISCOUNT_COST = 0;
-}
