--- src/main/java/christmas/util/constant/OrderManagerConst.java
@@ -1,8 +1,0 @@
-package christmas.util.constant;
-
-public class OrderManagerConst {
-    public static final String ORDER_DELIMITER = ",";
-    public static final String MENU_AND_AMOUNT_DELIMITER = "-";
-    public static final int INDEX_OF_MENU = 0;
-    public static final int INDEX_OF_AMOUNT = 1;
-}
