+++ src/main/java/christmas/validator/DomainValidator.java
@@ -1,7 +1,7 @@
 package christmas.validator;
 
-import static christmas.util.constant.ValidatorConst.COMPARE_EQUAL_VALUE_RESULT;
-import static christmas.util.constant.ValidatorConst.SINGLE_ORDER_CATEGORY;
+import static christmas.util.constant.ValidatorConstant.COMPARE_EQUAL_VALUE_RESULT;
+import static christmas.util.constant.ValidatorConstant.SINGLE_ORDER_CATEGORY;
 import static christmas.validator.ErrorMessage.ORDER_ERROR_MESSAGE;
 import static christmas.validator.ErrorMessage.PREFIX;
